+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+- Include markdown headings in TOC generation for the final page (not supported in editor TOC preview)
+
 ## [0.2.0-alpha.4] - 2026-04-19
 
 ### Added